firmware: website: moved website related strings to website folder
